src/assets/circle-k/requirements.txt
docxtpl==0.16.7
python-docx==0.8.11
pillow==9.5.0 